fixes bugs, refactors, adds photo and profile edit functionality alongside requisite redirects.  Need to still fix my-bookings route.
