+++ api/index.js
@@ -4,6 +4,7 @@ const mongoose = require("mongoose");
 const bcrypt = require('bcrypt');
 const jwt = require('jsonwebtoken');
 const User = require('./models/User.js');
+const Place = require('./models/Place.js');
 const cookieParser = require('cookie-parser');
 const imageDownloader = require('image-downloader');
 const multer = require('multer'); 
@@ -113,6 +114,57 @@ app.post('/upload', photosMiddleware.array('photos', 100), (req,res) => {
         uploadedFiles.push(newPath.replace('uploads/', ''));
     }
     res.json(uploadedFiles);
+});
+
+app.post('/places', (req,res) => {
+    const {token} = req.cookies;
+    const {
+        title,address,addedPhotos,description,
+        perks,extraInfo,checkIn,checkOut,maxGuests
+    } = req.body;
+    jwt.verify(token, jwtSecret, {}, async (err, userData) => {
+        if (err) throw err;
+        
+        const placeDoc = await Place.create({
+            owner: userData.id,
+            title,address,photos:addedPhotos,description,
+            perks,extraInfo,checkIn,checkOut,maxGuests
+        });
+        res.json(placeDoc)
+    });
+});
+
+app.get('/places', (req,res) => {
+    const {token} = req.cookies;
+    jwt.verify(token, jwtSecret, {}, async (err, userData) => {
+        const {id} = userData;
+        res.json( await Place.find({owner:id}) );
+    });
+});
+
+app.get('/places/:id', async (req,res) => {
+    const {id} = req.params;
+    res.json(await Place.findById(id));
+});
+
+app.put('/places/', async (req,res) => {
+    const {token} = req.cookies;
+    const {
+        id,title,address,addedPhotos,description,
+        perks,extraInfo,checkIn,checkOut,maxGuests
+    } = req.body;
+    jwt.verify(token, jwtSecret, {}, async (err, userData) => {
+        if (err) throw err;
+        const placeDoc = await Place.findById(id);
+        if (userData.id === placeDoc.owner.toString()) {
+            placeDoc.set({
+                title,address,photos:addedPhotos,description,
+                perks,extraInfo,checkIn,checkOut,maxGuests
+            });
+            await placeDoc.save();
+            res.json('ok');
+        }
+    });
 });
